third_party/blink/web_tests/platform/generic/accessibility/input-aria-required-expected.txt

This tests required and aria-required on input type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input1.isRequired is true
PASS input2.isRequired is true
PASS input3.isRequired is false
PASS input4.isRequired is true
PASS input5.isRequired is false
PASS successfullyParsed is true

TEST COMPLETE
